Module chat_flame_backend::llm::token_generator

Token generator utilities.

Provides the core functionality for generating individual tokens during the text generation process.

Modules

dummy

Structs

TokenGenerator

A token generator that generates tokens based on provided parameters, model processor, and sampler.

Enums

TokenGeneratorResult

Traits

TokenGeneratorTrait

A trait defining the behavior of a token generator.

Type Aliases

TokenProbability